requirements.txt
click==6.6
Flask==0.11.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
pygame==1.9.2a0
pyusb==1.0.0
virtualenv==15.0.3
Werkzeug==0.11.11
